requirements.txt
amqp==5.2.0
asgiref==3.8.1
async-timeout==4.0.3
autopep8==2.1.0
billiard==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.0.4
django-filter==24.2
django-redis==5.4.0
django-silk==5.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Faker==24.11.0
gprof2dot==2022.7.29
kombu==5.3.7
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycodestyle==2.11.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
toposort==1.10
typing_extensions==4.11.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
